backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
Pillow>=10.2.0
exifread>=3.0.0
deepface>=0.0.93
google-search-results>=2.4.2
httpx>=0.27.0
aiofiles>=23.2.1
